+++ frontend-ios/YakAl/YakAl/Sources/Models/MedicineModel.swift
@@ -1,10 +1,11 @@
-import Foundation
+import SwiftUI
 
-struct Medicine {
+struct Medicine:Identifiable {
     let id: Int
     let image: String
     let name: String
     let ingredients: String
     let dangerStat : Int
-    var isTaken: Bool
+    @State var isTaken: Bool //
+
 }
