+++ frontend-ios/YakAl/YakAl/Sources/VCs/Home/HomeVC.swift
@@ -1,31 +1,6 @@
 import UIKit
 import SwiftUI
 
-
-
-class MedicationData: ObservableObject {
-    @Published var medications: [Medication] = [
-//        Medication(name: "아침",medication: [
-//            Medicine(id: 1, image: "image_덱시로펜정", name: "덱시로펜정", ingredients: "해열, 진통, 소염제", dangerStat: 0, isTaken: false),
-//            Medicine(id: 2, image: "image_덱시로펜정", name: "동화디트로판정", ingredients: "소화성 궤양용제", dangerStat: 1, isTaken: false),
-//            Medicine(id: 3, image: "image_덱시로펜정", name: "동광레바미피드정", ingredients: "소화성 궤양용제", dangerStat: 1, isTaken: false)
-//        ]),
-//        Medication(name: "점심",medication: [
-//            Medicine(id: 3, image: "image_덱시로펜정", name: "약물C", ingredients: "성분 C", dangerStat: 2, isTaken: false),
-//            Medicine(id: 4, image: "image_덱시로펜정", name: "약물D", ingredients: "성분 D", dangerStat: 1, isTaken: false)
-//        ]),
-//        Medication(name: "저녁", medication: [
-//            Medicine(id: 5, image: "image_덱시로펜정", name: "약물E", ingredients: "성분 E", dangerStat: 2, isTaken: false),
-//            Medicine(id: 6, image: "image_덱시로펜정", name: "약물F", ingredients: "성분 F", dangerStat: 1, isTaken: false)
-//        ]),
-//        Medication(name: "기타",medication: [
-//            Medicine(id: 7, image: "image_덱시로펜정", name: "약물G", ingredients: "성분 G", dangerStat: 0, isTaken: false),
-//            Medicine(id: 8, image: "image_덱시로펜정", name: "약물H", ingredients: "성분 H", dangerStat: 1, isTaken: false),
-//            Medicine(id: 9, image: "image_덱시로펜정", name: "약물J", ingredients: "성분 J", dangerStat: 2, isTaken: false)
-//        ]),
-    ]
-}
-
 class HomeVC: UIViewController{
     // MARK: - Outlets -
     @IBOutlet weak var calendarView: UIView!
@@ -45,6 +20,10 @@ class HomeVC: UIViewController{
     
     @IBOutlet weak var addMedicineButton: UIButton!
     
+    @IBOutlet weak var todayMedicinCnt: UILabel!
+    
+    
+    
     
     // MARK: - Properties
     let progressCircle = CAShapeLayer()
@@ -62,6 +41,7 @@ class HomeVC: UIViewController{
     
     
     var medicationData = MedicationData()
+    
     var medicationHostingController: UIHostingController<MedicationSwiftUIView>?
 
 
@@ -87,14 +67,7 @@ class HomeVC: UIViewController{
         //MARK: - viewDidLoad
         override func viewDidLoad() {
             super.viewDidLoad()
-            
-            
-            NSLayoutConstraint.activate([
-                emptyView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-                emptyView.topAnchor.constraint(equalTo: view.topAnchor),
-                emptyView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-                emptyView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
-            ])
+            todayMedicinCnt.text = "총 \(medicationData.totalMedicineCount)개"
 
             
             if medicationData.medications.isEmpty {
@@ -224,6 +197,7 @@ class HomeVC: UIViewController{
         @IBAction func AlertButtonAction(_ sender: UIButton) {
             print("Alert")
             animateProgress()
+            print(medicationData.totalCompletedCount)
         }
         @IBAction func EnvelopeMedicineButtonAction(_ sender: UIButton) {
             print("EnvelopeMedicineButtonAction")
